科室、挂号删除细节处理

+++ src/main/java/com/hyit/hospital/service/Impl/DepartmentServiceImpl.java
@@ -4,13 +4,17 @@ import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import com.hyit.hospital.common.Common;
 import com.hyit.hospital.mapper.DepartmentMapper;
+import com.hyit.hospital.mapper.UserMapper;
 import com.hyit.hospital.model.Department;
+import com.hyit.hospital.model.User;
 import com.hyit.hospital.service.DepartmentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.DeleteMapping;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @author hc
@@ -21,6 +25,9 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Autowired
     private DepartmentMapper departmentMapper;
+
+    @Autowired
+    private UserMapper userMapper;
 
     @Override
     public List<Department> selAll() {
@@ -42,9 +49,24 @@ public class DepartmentServiceImpl implements DepartmentService {
         departmentMapper.insert(department);
     }
 
+    //删除前判断是否科室绑定用户
     @Override
-    public void delById(Integer id) {
-        departmentMapper.deleteByPrimaryKey(id);
+    public Map<String, Object> delById(Integer id) {
+        Map<String,Object> map = new HashMap<>();
+        User user = new User();
+        user.setDeptid(id);
+        List<User> users = userMapper.selAll(user);
+        if(users.isEmpty()){
+            //成功
+            map.put("status",1);
+            departmentMapper.deleteByPrimaryKey(id);
+            return map;
+
+        }else {
+            //不成功
+            map.put("status",0);
+            return map;
+        }
     }
 
     @Override
