+++ src/main/java/com/hyit/hospital/service/Impl/RegistlevelServiceImpl.java
@@ -4,12 +4,16 @@ import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import com.hyit.hospital.common.Common;
 import com.hyit.hospital.mapper.RegistlevelMapper;
+import com.hyit.hospital.mapper.UserMapper;
 import com.hyit.hospital.model.Registlevel;
+import com.hyit.hospital.model.User;
 import com.hyit.hospital.service.RegistlevelService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @author hc
@@ -20,6 +24,9 @@ public class RegistlevelServiceImpl implements RegistlevelService {
 
     @Autowired
     private RegistlevelMapper registlevelMapper;
+
+    @Autowired
+    private UserMapper userMapper;
 
     @Override
     public List<Registlevel> selAll() {
@@ -37,12 +44,29 @@ public class RegistlevelServiceImpl implements RegistlevelService {
 
     @Override
     public void registlevelAdd(Registlevel registlevel) {
+        registlevel.setDelmark(1);
         registlevelMapper.insert(registlevel);
     }
 
+    //删除前判断是否科室绑定用户
     @Override
-    public void delById(Integer id) {
-        registlevelMapper.deleteByPrimaryKey(id);
+    public Map<String,Object> delById(Integer id) {
+
+        Map<String,Object> map = new HashMap<>();
+        User user = new User();
+        user.setRegistleid(id);
+        List<User> users = userMapper.selAll(user);
+        if(users.isEmpty()){
+            //成功
+            map.put("status",1);
+            registlevelMapper.deleteByPrimaryKey(id);
+            return map;
+
+        }else {
+            //不成功
+            map.put("status",0);
+            return map;
+        }
     }
 
     @Override
